# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Setlists API",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "description": "Attempts to validate the user with a password. If so, sends back the user name and a personal token.",
                "operationId": "App\\Http\\Controllers\\AuthController::login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "The email of the user that wants to use the API.",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "The password of the user that wants to use the API.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "A json containing the name of the user and a token. This token must be appended to the url of every request in this form: http://example.com/api/songs?token=[token].\n     * The token is every time this request is done and it will last for 1 hour.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Error: invalid username or password."
                    },
                    "400": {
                        "description": "Error: invalid request."
                    }
                }
            }
        },
        "/api/auth/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "description": "Removes the token for the current user. This route requires a valid token in the url",
                "operationId": "App\\Http\\Controllers\\AuthController::logout",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "A valid and recent token.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A success message indicating that the token has been removed from the system, thus forcing the user to create a new one in order to kkep using the API.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/expired"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "responses": {
            "expired": {
                "description": "Error: the provided token has expired, or is invalid or is obsolete."
            },
            "unauthorized": {
                "description": "Error: unauthorized request (no token present)"
            },
...
        }
    },
...
}